requirements.txt
# Multi-Model Algorithmic Trading Platform
# Python 3.9+ Required

# Core Data Processing
pandas>=2.1.0,<3.0.0
numpy>=1.25.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Data Storage
pyarrow>=13.0.0  # Parquet support

# Database
duckdb>=0.9.0

# Configuration & Validation
pyyaml>=6.0.1
pydantic>=2.4.0,<3.0.0

# Broker APIs
alpaca-py>=0.12.0  # Alpaca trading (equities)
ccxt>=4.3.0  # Multi-exchange crypto trading (using latest stable)

# Data Sources
yfinance>=0.2.28  # Yahoo Finance historical data

# Optimization
deap>=1.4.1  # Evolutionary algorithms

# Logging
python-json-logger>=2.0.7

# Visualization
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools (optional)
black>=23.9.1
flake8>=6.1.0
mypy>=1.5.1
